more typescript ports
